refactor useFleetPrometheusPoll into a file
