+++ frontend/packages/multicluster-sdk/src/internal/usePoll.ts
