+++ frontend/packages/multicluster-sdk/src/internal/useSafeFetch.ts
